+++ src/main/kotlin/ru/raysmith/tgbot/model/network/media/Media.kt
@@ -1,8 +1,6 @@
 package ru.raysmith.tgbot.model.network.media
 
 import ru.raysmith.tgbot.core.ApiCaller
-import ru.raysmith.tgbot.utils.errorBody
-import java.io.InputStream
 
 /** This interface represents media attachment */
 interface Media {
@@ -13,8 +11,8 @@ interface Media {
     /** Unique identifier for this file, which is supposed to be the same over time and for different bots. Can't be used to download or reuse the file. */
     val fileUniqueId: String
 
-    /** File size */
-    val fileSize: Int?
+    /** File size in bytes */
+    val fileSize: Long?
 
     /** Original animation filename as defined by sender */
     val fileName: String?
