+++ src/main/kotlin/ru/raysmith/tgbot/model/network/media/PhotoSize.kt
@@ -9,7 +9,7 @@ data class PhotoSize(
     @SerialName("file_unique_id") override val fileUniqueId: String,
     @SerialName("width") val width: Int,
     @SerialName("height") val height: Int,
-    @SerialName("file_size") override val fileSize: Int? = null,
+    @SerialName("file_size") override val fileSize: Long? = null,
 ) : Media {
     override val fileName: String? = null
 }
