+++ src/main/kotlin/ru/raysmith/tgbot/model/network/media/Voice.kt
@@ -8,6 +8,6 @@ data class Voice(
     @SerialName("file_id") override val fileId: String,
     @SerialName("file_unique_id") override val fileUniqueId: String,
     @SerialName("file_name") override val fileName: String? = null,
-    @SerialName("file_size") override val fileSize: Int? = null,
+    @SerialName("file_size") override val fileSize: Long? = null,
     @SerialName("duration") val duration: Int,
 ) : Media
